+++ front/package.json
@@ -9,12 +9,16 @@
     "eject": "react-scripts eject"
   },
   "dependencies": {
+    "@material-ui/core": "^4.9.5",
+    "@material-ui/lab": "^4.0.0-alpha.45",
     "@testing-library/jest-dom": "^4.2.4",
     "@testing-library/react": "^9.3.2",
     "@testing-library/user-event": "^7.1.2",
     "bootstrap": "^4.4.1",
+    "material-ui": "^0.20.2",
     "react": "^16.13.0",
     "react-bootstrap": "^1.0.0-beta.17",
+    "react-bootstrap-typeahead": "^4.0.0",
     "react-calendar": "^2.19.2",
     "react-cool-onclickoutside": "^1.3.2",
     "react-dom": "^16.13.0",
