--- front/src/components/HomePage/ReactAuto.jsx
@@ -1,7 +1,0 @@
-import React from 'react';
-
-function ReactAutoComplete() {
-    return <div></div>;
-}
-
-export default ReactAutoComplete;
